# openapi.yaml
openapi: 3.1.0
info:
  title: Fluxora HTTP API
  version: 0.1.0
  description: |
    Fluxora Backend: Off-chain companion to the streaming contract.

    Presents operator-grade reliability with predictable HTTP semantics, durable views of chain-derived data,
    and explicit failure behavior when dependencies are unhealthy.

    ## Security Header Baseline
    Fluxora applies Helmet at the Express boundary so every HTTP response carries a consistent
    browser-facing security-header baseline, including `Content-Security-Policy`,
    `Strict-Transport-Security`, `X-Content-Type-Options`, `X-Frame-Options`, and
    `Referrer-Policy`. The service also suppresses the `X-Powered-By` header.

    These headers are transport invariants for public clients, authenticated partners,
    administrators, and internal workers that traverse the HTTP surface. They do not replace
    authentication, authorization, TLS termination, or proxy hardening.

    ## Trust Boundaries
    - **Public Internet Clients**: Read-only access to stream listings and health checks
    - **Authenticated Partners**: Create and manage streams with idempotency guarantees
    - **Administrators**: Full access including internal indexer endpoints
    - **Internal Workers**: Indexer endpoints for chain state synchronization

    ## Failure Modes & Client-Visible Behavior
    - **Invalid Input (400)**: Malformed addresses, negative amounts, oversized payloads
...

  contact:
    name: Fluxora Team
  license:
...
servers:
...
paths:
...
  /api/auth/session:
    post:
      summary: Create authentication session
      operationId: createSession
      tags:
        - Auth
      description: |
        Issues a JWT token for authenticated access to privileged endpoints.
        In this iteration, any valid Stellar address can obtain a session.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - address
              properties:
                address:
                  type: string
                  description: Stellar public key
                  pattern: '^G[A-Z2-7]{55}$'
                  example: GBBD47UZQ5CYVVEUVRYNQZX3G5KRZTAYF5XSVS2UKMCCWW5LJJLXNVQX
                role:
                  type: string
                  enum: [operator, viewer]
                  default: viewer
                  description: User role for authorization
      responses:
        '200':
          description: Session created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    ResponseMeta:
      type: object
      required:
        - timestamp
      properties:
        timestamp:
          type: string
          format: date-time
          description: ISO-8601 timestamp of the response
          example: '2024-01-01T12:00:00.000Z'
        requestId:
          type: string
          description: Opaque request identifier for log correlation
          example: 550e8400-e29b-41d4-a716-446655440000
...
    SessionResponse:
      type: object
      required:
        - success
        - data
        - meta
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          type: object
          properties:
            token:
              type: string
              description: JWT token for authentication
              example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
            user:
              type: object
              properties:
                address:
                  type: string
                  example: GBBD47UZQ5CYVVEUVRYNQZX3G5KRZTAYF5XSVS2UKMCCWW5LJJLXNVQX
                role:
                  type: string
                  example: operator
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          enum: [false]
          description: Always false for error responses
          example: false
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Machine-readable error code
              enum:
                - VALIDATION_ERROR
                - NOT_FOUND
                - CONFLICT
                - UNAUTHORIZED
                - FORBIDDEN
                - PAYLOAD_TOO_LARGE
                - TOO_MANY_REQUESTS
                - REQUEST_TIMEOUT
                - SERVICE_UNAVAILABLE
                - INTERNAL_ERROR
                - DECIMAL_ERROR
                - UNSUPPORTED_MEDIA_TYPE
                - GATEWAY_TIMEOUT
                - NOT_ACCEPTABLE
                - PERSISTED_QUERY_INVALID
                - PERSISTED_QUERY_UNSUPPORTED_VERSION
                - PERSISTED_QUERY_INVALID_HASH
                - PERSISTED_QUERY_HASH_MISMATCH
                - PERSISTED_QUERY_NOT_FOUND
                - GRAPHQL_PARSE_ERROR
                - FEATURE_FLAG_DISABLED
                - CONFIGURATION_ERROR
                - RESTORE_ERROR
                - PERSISTENCE_ERROR
                - API_KEY_ERROR
                - DIAGNOSTICS_ERROR
                - SERVICE_SHUTTING_DOWN
                - HEALTH_CHECK_ERROR
                - CONSUMER_SUSPENDED
                - ENTRY_ALREADY_REPLAYED
                - INVALID_REQUEST
                - QUEUE_ERROR
                - INVALID_DELIVERY_ID
                - DELIVERY_NOT_FOUND
                - INVALID_PAGINATION
                - INVALID_OUTBOX_FILTER
                - INVALID_RETRY_REQUEST
                - DLQ_ITEM_NOT_FOUND
                - DLQ_PROCESS_ERROR
                - DLQ_RETRY_ERROR
                - INVALID_ENDPOINT_URL
                - OUTBOX_PROCESSING_ERROR
                - INVALID_CLEANUP_REQUEST
                - CLEANUP_ERROR
                - CORS_ORIGIN_DENIED
                - RATE_LIMIT_EXCEEDED
                - INVALID_ADDRESS
                - ERASURE_FAILED
                - INTERNAL_CALLER_ADDRESS_MISSING
                - unsupported_media_type
                - payload_too_large
                - invalid_encoding
                - payload_too_deep
                - invalid_json
                - missing_secret
                - missing_delivery_id
                - missing_timestamp
                - missing_signature
                - invalid_timestamp
                - timestamp_outside_tolerance
                - previous_secret_expired
                - signature_mismatch
                - duplicate_delivery
                - UNSUPPORTED_VERSION
                - IDEMPOTENCY_CONFLICT
              example: VALIDATION_ERROR
            message:
              type: string
              description: Human-readable error message
              example: Validation failed
            requestId:
              type: string
              description: Correlation ID for log triage
              example: 550e8400-e29b-41d4-a716-446655440000
            details:
              description: Additional error context (field-level validation errors, etc.)
              nullable: true
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token for authentication. Include in Authorization header:
        Authorization: Bearer <token>
    AdminAuth:
      type: http
      scheme: bearer
      description: |
        Admin API key for privileged operations. Include in Authorization header:
        Authorization: Bearer <ADMIN_API_KEY>
        Set via ADMIN_API_KEY environment variable.
...